pack cast propagation
